// models/tour.model.ts
export interface TourModel {
  id: string;

  title_en: string;
  title_ru: string;
  title_tr: string;

  description_en: string;
  description_ru: string;
  description_tr: string;

  price_en: string;
  price_ru: string;
  price_tr: string;

  imagePath: string;

  createdAt: string;
  updatedAt: string;
}